+++ server/package.json
@@ -11,7 +11,9 @@
   "author": "shashi",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.19.2"
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "mongoose": "^8.4.3"
   },
   "devDependencies": {
     "nodemon": "^3.1.3"
